Refactor: Move the WebGPU initialization module
While it isn't exactly worth putting this into the runtime, it doesn't really belong to the client or even the renderer proper.
